# src/swagger.yaml
openapi: 3.0.0
info:
  title: Pet API
  version: 1.0.0
  description: A simple API to manage pets.
servers:
  - url: http://localhost:3000
paths:
  /api/pets:
    get:
      summary: Get a list of pets
      responses:
        '200':
          description: A list of pets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
    post:
      summary: Create a new pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '201':
          description: Pet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
  /api/pets/{id}:
    get:
      summary: Get a pet by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the pet to get
          schema:
            type: integer
      responses:
        '200':
          description: A single pet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found
    put:
      summary: Update a pet by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the pet to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '200':
          description: Pet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found
    delete:
      summary: Delete a pet by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the pet to delete
          schema:
            type: integer
      responses:
        '200':
          description: Pet deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found
components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Buddy'
        type:
          type: string
          example: 'Dog'